% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_syn_data.R
\name{generate_synthetic_visium}
\alias{generate_synthetic_visium}
\title{Generate synthetic Visium/ST data from input scRNAseq data.}
\usage{
generate_synthetic_visium(seurat_obj, dataset_type, clust_var, n_regions, region_var = NULL, dataset_id = "1", n_spots_min = 50, n_spots_max = 500, visium_mean = 20000, visium_sd = 7000, n_cells_min = 2, n_cells_max = 10, add_mock_region = FALSE, sc_rnaseq_path = NA)
}
\arguments{
\item{seurat_obj}{The input scRNAseq data stored as Seurat object.}

\item{dataset_type}{Name indicating which type of synthetic Visium data needs to be generated. \cr \cr
"real": for each region (defined in scRNAseq data - region_var - and thus 'real'), cell type frequencies are based on the number of cells in the corresponding region in the input scRNAseq data \cr \cr
"real_top1": for each region (defined in scRNAseq data and thus 'real'), we will only consider cell types that are most frequently present in that 1 region ("top 1"). Cell type frequencies are based on the number of cells in the corresponding region in the input scRNAseq data \cr \cr
"real_top1_uniform": for each region (defined in scRNAseq data - region_var - and thus 'real'), we will only consider cell types that are most frequently present in that 1 region ("top 1"). Cell type frequencies are equal for each cell type assigned to a region ("uniform") \cr \cr
"real_top2_overlap": for each region (defined in scRNAseq data - region_var - and thus 'real'), we will only consider cell types that if this particular region is among the 2 regions in which they are most abundant ("top 2"). As a consequence, there will be some "overlap" in cell type presence between two regions.  Cell type frequencies are based on the number of cells in the corresponding region in the input scRNAseq data \cr \cr
"real_top2_overlap_uniform": for each region (defined in scRNAseq data - region_var - and thus 'real'), we will only consider cell types that if this particular region is among the 2 regions in which they are most abundant ("top 2"). As a consequence, there will be some "overlap" in cell type presence between two regions. Cell type frequencies are equal for each cell type assigned to a region ("uniform") \cr \cr
"real_missing_celltypes_visium": the same as the 'real' dataset type, but now some random input scRNAseq cell types will be left out, and thus not used to generate synthetic data. \cr \cr
"artificial_uniform_distinct": celltype-prior region assignments are entirely artificial and thus not based on possibly present regional information in the input scRNAseq data. Each artifical prior region has a distinct set of cell types here, and cell type frequencies are the same for each cell type in each region. \cr \cr
"artificial_diverse_distinct": celltype-prior region assignments are entirely artificial and thus not based on possibly present regional information in the input scRNAseq data. Each artifical prior region has a distinct set of cell types here, and cell type frequencies are different for each cell type in each region.  \cr \cr
"artificial_uniform_overlap": celltype-prior region assignments are entirely artificial and thus not based on possibly present regional information in the input scRNAseq data. Each artifical prior region could have some overlap in cell type contribution with other regions, and cell type frequencies are the same for each cell type in each region.  \cr \cr
"artificial_diverse_overlap": celltype-prior region assignments are entirely artificial and thus not based on possibly present regional information in the input scRNAseq data. Each artifical prior region could have some overlap in cell type contribution with other regions, and cell type frequencies are different for each cell type in each region.  \cr \cr
"artificial_dominant_celltype_diverse": celltype-prior region assignments are entirely artificial and thus not based on possibly present regional information in the input scRNAseq data. Each artifical prior region could have some overlap in cell type contribution with other regions, and cell type frequencies are different for each cell type in each region. There is one dominant celltype that is present in each region, and is much more abundant than other cell types (5-15x more). \cr \cr
"artificial_partially_dominant_celltype_diverse: celltype-prior region assignments are entirely artificial and thus not based on possibly present regional information in the input scRNAseq data. Each artifical prior region could have some overlap in cell type contribution with other regions, and cell type frequencies are different for each cell type in each region. There is one dominant celltype that is present in all but one regions ("partially dominant"), and is much more abundant than other cell types, except for one region where it is equally abundant. \cr \cr
"artificial_missing_celltypes_visium": the same as the 'artificial_diverse_overlap' dataset type, but now some random input scRNAseq cell types will be left out, and thus not used to generate synthetic data. \cr \cr
"artificial_dominant_rare_celltype_diverse": the same as 'artificial_dominant_celltype_diverse', except that the dominant cell type that is present in all regions is much less abundant than other cell types. \cr \cr
"artificial_regional_rare_celltype_diverse": the same as 'artificial_dominant_rare_celltype_diverse', except that the rare cell type is now only present in one region instead of all regions.\cr \cr
"artificial_diverse_distinct_missing_celltype_sc": the same as 'artificial_diverse_distinct' except that one random cell type will be removed from the reference scRNAseq dataset at time of integration and evaluation. This to resemble a case where integration is done with an incomplete reference. \cr \cr
"artificial_diverse_overlap_missing_celltype_sc": the same as 'artificial_diverse_overlap' except that one random cell type will be removed from the reference scRNAseq dataset at time of integration and evaluation. This to resemble a case where integration is done with an incomplete reference. \cr \cr}

\item{clust_var}{Name of the meta data column in which the cell type label of interest is defined.}

\item{n_regions}{Number of regions you want to define. If applicable: this can be NULL if you want to use the same number of regions as defined in a meta data column indicating the region of interest.}

\item{region_var}{Default NULL. If your scRNAseq data object already provides information about the region/location a cell was sampled from, you could change this argument to the meta data column indicating the region of interest.}

\item{dataset_id}{Character vector to define how you want to identify the dataset. The final identifier/name of the dataset will be the description of the dataset type followed by this character vector.}

\item{n_spots_min}{The minimum number of spots allowed per region. Default 50.}

\item{n_spots_max}{The maximum number of spots allowed per region. Default: 500.}

\item{visium_mean}{The mean of the normal distribution that will be used to pick the target number of counts for downsample.}

\item{visium_sd}{The standard deviation of the normal distribution that will be used to pick the target number of counts for downsample.}

\item{n_cells_min}{Integer indicating the minimum number of cells sampled per spot. Default: 2}

\item{n_cells_max}{Integer indicating the maximum number of cells sampled per spot. Default: 10}

\item{add_mock_region}{If you want to make a "mock region" in addition to the other regions, set this parameter to TRUE. Mock region is recommended when the synthetic data will be used to evaluate spatial/region annotation of cells, not when evaluating deconvolution tools. Default: FALSE. The mock region is generated similar as the real regions, but the input cell type frequencies are the same for each cell type, and after generation of the counts, the gene names are shuffled such that cellular identities in this mockregion are lost.}

\item{sc_rnaseq_path}{NA or indication of the path to read in the scRNAseq that should be used to integratie with this synthetic visium data. Default: NA}
}
\value{
list with six sublists: \cr
counts: count matrix of the synthetic visium data \cr
spot_composition: the cell type composition of each spot \cr
relative_spot_composition: the relative cell type composition of each spot \cr
gold_standard_priorregion: indication of which cell types belong to which prior region \cr
dataset_properties: indication of the properties of the dataset
sc_rnaseq_path: NA or indication of the path to read in the scRNAseq that should be used to integratie with this synthetic visium data
}
\description{
\code{generate_synthetic_visium} This function generates synthetic Visium/ST data from input scRNAseq data. \cr
To create synthetic Visium spots, we will generate an artificial mixture of single cells (based on the assumption that a spot is a mixture of n cells \cr
Cells from the input scRNAseq data will be sampled based on the provided cell type frequencies per region (such that each region has a different cellular composition), and their counts will be summed. \cr
After summation of the counts, the counts will be downsampled to get a total number of counts per spot that is typical for a Visium dataset. This target number of counts will be determined by sampling from a Normal distribution with mean and SD provided by the user. \cr
You can generate different 'tissue types' by changing the way the cell type frequencies are defined. Read the argument description of `dataset_type` for seeing the different tissue type options. \cr
If your scRNAseq data object already provides information about the region/location a cell was sampled from, you could use this information as well to generate synthetic Visium data that will only pool cells together if they originate from the same region. \cr
}
\examples{
\dontrun{
library(Seurat)
synthetic_visium_data = generate_synthetic_visium(seurat_obj = seurat_obj, dataset_type = "artificial_missing_celltypes_visium", clust_var = "subclass", n_regions = 5)
}

}
