% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_syn_data.R
\name{generate_synthetic_visium_lite}
\alias{generate_synthetic_visium_lite}
\title{Generate synthetic Visium/ST data from input scRNAseq data, using the cell type composition from the input as priors.}
\usage{
generate_synthetic_visium_lite(seurat_obj, clust_var, dataset_id = "1", n_spots = 500, visium_mean = 20000, visium_sd = 7000, n_cells_min = 2, n_cells_max = 10, sc_rnaseq_path = NA)
}
\arguments{
\item{seurat_obj}{The input scRNAseq data stored as Seurat object.}

\item{clust_var}{Name of the meta data column in which the cell type label of interest is defined.}

\item{dataset_id}{Character vector to define how you want to identify the dataset. The final identifier/name of the dataset will be the description of the dataset type followed by this character vector.}

\item{n_spots}{The number of spots to be generated. Default 500.}

\item{visium_mean}{The mean of the normal distribution that will be used to pick the target number of counts for downsample.}

\item{visium_sd}{The standard deviation of the normal distribution that will be used to pick the target number of counts for downsample.}

\item{n_cells_min}{Integer indicating the minimum number of cells sampled per spot. Default: 2}

\item{n_cells_max}{Integer indicating the maximum number of cells sampled per spot. Default: 10}

\item{sc_rnaseq_path}{NA or indication of the path to read in the scRNAseq that should be used to integratie with this synthetic visium data. Default: NA}
}
\value{
list with six sublists: \cr
counts: count matrix of the synthetic visium data \cr
spot_composition: the cell type composition of each spot \cr
relative_spot_composition: the relative cell type composition of each spot \cr
gold_standard_priorregion: indication of which cell types belong to which prior region \cr
dataset_properties: indication of the properties of the dataset
sc_rnaseq_path: NA or indication of the path to read in the scRNAseq that should be used to integratie with this synthetic visium data
}
\description{
\code{generate_synthetic_visium_lite} This function generates synthetic Visium/ST data from input scRNAseq data. \cr
In contrast to the original function, the cell type frequencies are based on the cell type composition of the input scRNAseq data. \cr
Because of this, there can only be 1 region. It is recommended to use single-nucleus RNA-seq data, as it has been shown to capture tissue compositions more accurately than scRNA-seq. \cr
}
\examples{
\dontrun{
library(Seurat)
synthetic_visium_data = generate_synthetic_visium_lite(seurat_obj = seurat_obj, clust_var = "subclass")
}

}
