% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_syn_data.R
\name{make_region_celltype_assignment}
\alias{make_region_celltype_assignment}
\title{Prepare the generation of synthetic Visium/ST data by using a scRNAseq Seurat object as input and defining a priori defined cell type frequencies per region (group of spots).}
\usage{
make_region_celltype_assignment(seurat_obj, clust_var, n_regions, dataset_type, region_var = NULL, dataset_id = "1", n_spots_min = 50, n_spots_max = 500)
}
\arguments{
\item{seurat_obj}{The input scRNAseq data stored as Seurat object.}

\item{clust_var}{Name of the meta data column in which the cell type label of interest is defined.}

\item{n_regions}{Number of regions you want to define. If applicable: this can be NULL if you want to use the same number of regions as defined in a meta data column indicating the region of interest.}

\item{dataset_type}{Name indicating which type of synthetic Visium data needs to be generated. \cr \cr
"real": for each region (defined in scRNAseq data - region_var - and thus 'real'), cell type frequencies are based on the number of cells in the corresponding region in the input scRNAseq data \cr \cr
"real_top1": for each region (defined in scRNAseq data and thus 'real'), we will only consider cell types that are most frequently present in that 1 region ("top 1"). Cell type frequencies are based on the number of cells in the corresponding region in the input scRNAseq data \cr \cr
"real_top1_uniform": for each region (defined in scRNAseq data - region_var - and thus 'real'), we will only consider cell types that are most frequently present in that 1 region ("top 1"). Cell type frequencies are equal for each cell type assigned to a region ("uniform") \cr \cr
"real_top2_overlap": for each region (defined in scRNAseq data - region_var - and thus 'real'), we will only consider cell types that if this particular region is among the 2 regions in which they are most abundant ("top 2"). As a consequence, there will be some "overlap" in cell type presence between two regions.  Cell type frequencies are based on the number of cells in the corresponding region in the input scRNAseq data \cr \cr
"real_top2_overlap_uniform": for each region (defined in scRNAseq data - region_var - and thus 'real'), we will only consider cell types that if this particular region is among the 2 regions in which they are most abundant ("top 2"). As a consequence, there will be some "overlap" in cell type presence between two regions. Cell type frequencies are equal for each cell type assigned to a region ("uniform") \cr \cr
"real_missing_celltypes_visium": the same as the 'real' dataset type, but now some random input scRNAseq cell types will be left out, and thus not used to generate synthetic data. \cr \cr
"artificial_uniform_distinct": celltype-prior region assignments are entirely artificial and thus not based on possibly present regional information in the input scRNAseq data. Each artifical prior region has a distinct set of cell types here, and cell type frequencies are the same for each cell type in each region. \cr \cr
"artificial_diverse_distinct": celltype-prior region assignments are entirely artificial and thus not based on possibly present regional information in the input scRNAseq data. Each artifical prior region has a distinct set of cell types here, and cell type frequencies are different for each cell type in each region.  \cr \cr
"artificial_uniform_overlap": celltype-prior region assignments are entirely artificial and thus not based on possibly present regional information in the input scRNAseq data. Each artifical prior region could have some overlap in cell type contribution with other regions, and cell type frequencies are the same for each cell type in each region.  \cr \cr
"artificial_diverse_overlap": celltype-prior region assignments are entirely artificial and thus not based on possibly present regional information in the input scRNAseq data. Each artifical prior region could have some overlap in cell type contribution with other regions, and cell type frequencies are different for each cell type in each region.  \cr \cr
"artificial_dominant_celltype_diverse": celltype-prior region assignments are entirely artificial and thus not based on possibly present regional information in the input scRNAseq data. Each artifical prior region could have some overlap in cell type contribution with other regions, and cell type frequencies are different for each cell type in each region. There is one dominant celltype that is present in each region, and is much more abundant than other cell types (5-15x more). \cr \cr
"artificial_partially_dominant_celltype_diverse: celltype-prior region assignments are entirely artificial and thus not based on possibly present regional information in the input scRNAseq data. Each artifical prior region could have some overlap in cell type contribution with other regions, and cell type frequencies are different for each cell type in each region. There is one dominant celltype that is present in all but one regions ("partially dominant"), and is much more abundant than other cell types, except for one region where it is equally abundant. \cr \cr
"artificial_missing_celltypes_visium": the same as the 'artificial_diverse_overlap' dataset type, but now some random input scRNAseq cell types will be left out, and thus not used to generate synthetic data. \cr \cr}

\item{region_var}{Default NULL. If your scRNAseq data object already provides information about the region/location a cell was sampled from, you could change this argument to the meta data column indicating the region of interest.}

\item{dataset_id}{Character vector to define how you want to identify the dataset. The final identifier/name of the dataset will be the description of the dataset type followed by this character vector.}

\item{n_spots_min}{The minimum number of spots allowed per region. Default 50.}

\item{n_spots_max}{The maximum number of spots allowed per region. Default: 500.}
}
\value{
A list with three sublists: \cr
region_assignments: gives the number of spots and cell type frequencies that will be used for each prior region \cr
dataset_properties: a tibble describing the dataset properties (based on the dataset_type argument) \cr
gold_standard_priorregion: a tibble indicating the presence and frequency of each scRNAseq-derived cell type in each prior region
}
\description{
\code{make_region_celltype_assignment} will start from input scRNAseq data (Seurat object) and generate cell type frequencies that are used as base for generating synthetic Visium.
Via this functions you can generate different 'tissue types' by changing the way the cell type frequencies are defined. Read the argument description of `dataset_type` for seeing the different tissue type options.
If your scRNAseq data object already provides information about the region/location a cell was sampled from, you could use this information as well to generate synthetic Visium data that will only pool cells together if they originate from the same region.
}
\examples{
\dontrun{
library(Seurat)
region_assignment = make_region_celltype_assignment(seurat_obj = seurat_obj, clust_var = "subclass", n_regions = 5, dataset_type = "artificial_missing_celltypes_visium")
}

}
