% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_syn_data.R
\name{region_assignment_to_syn_data}
\alias{region_assignment_to_syn_data}
\title{Generate synthetic Visium/ST data from input scRNAseq data based on predefined cell type frequencies and number of spots.}
\usage{
region_assignment_to_syn_data(region_assignment_list, seurat_obj, clust_var, visium_mean = 20000, visium_sd = 7000, n_cells_min = 2, n_cells_max = 10, add_mock_region = FALSE)
}
\arguments{
\item{region_assignment_list}{Output of the function `make_region_celltype_assignment`}

\item{seurat_obj}{The input scRNAseq data stored as Seurat object.}

\item{clust_var}{Name of the meta data column in which the cell type label of interest is defined.}

\item{visium_mean}{The mean of the normal distribution that will be used to pick the target number of counts for downsample.}

\item{visium_sd}{The standard deviation of the normal distribution that will be used to pick the target number of counts for downsample.}

\item{n_cells_min}{Integer indicating the minimum number of cells sampled per spot. Default: 2}

\item{n_cells_max}{Integer indicating the maximum number of cells sampled per spot. Default: 10}

\item{add_mock_region}{If you want to make a "mock region" in addition to the other regions, set this parameter to TRUE. Mock region is recommended when the synthetic data will be used to evaluate spatial/region annotation of cells, not when evaluating deconvolution tools. Default: FALSE. The mock region is generated similar as the real regions, but the input cell type frequencies are the same for each cell type, and after generation of the counts, the gene names are shuffled such that cellular identities in this mockregion are lost.}
}
\value{
list with five sublists: \cr
counts: count matrix of the synthetic visium data \cr
spot_composition: the cell type composition of each spot \cr
relative_spot_composition: the relative cell type composition of each spot \cr
gold_standard_priorregion: indication of which cell types belong to which prior region \cr
dataset_properties: indication of the properties of the dataset (see `make_region_celltype_assignment`)
}
\description{
\code{region_assignment_to_syn_data} Generate synthetic Visium/ST data based on predefined cell type frequencies and number of spots.
To create synthetic Visium spots, we will generate an artificial mixture of single cells (based on the assumption that a spot is a mixture of n cells (n between 2 and 10))
Cells from the input scRNAseq data will be sampled based on the provided cell type frequencies per region (such that each region has a different cellular composition), and their counts will be summed.
After summation of the counts, the counts will be downsampled to get a total number of counts per spot that is typical for a Visium dataset. This target number of counts will be determined by sampling from a Normal distribution with mean and SD provided by the user.
}
\examples{
\dontrun{
library(Seurat)
region_assignment = make_region_celltype_assignment(seurat_obj = seurat_obj, clust_var = "subclass", n_regions = 5, dataset_type = "artificial_missing_celltypes_visium")
synthetic_visium_data = region_assignment_to_syn_data(region_assignment_list = region_assignment, seurat_obj = seurat_obj, clust_var = "subclass")
}

}
