% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seurat_obj}
\alias{seurat_obj}
\title{Toy Seurat Object with subsampled cells and features.}
\format{
A Seurat Object (check https://satijalab.org/seurat/)
}
\usage{
seurat_obj
}
\description{
Toy Seurat Object with subsampled cells and features, based on the Allen Brain Atlas reference dataset.
}
\keyword{datasets}
